Honeybee Nest Ventilation Model
This function contains a user friendly simulation of ventilation dynamics in honeybee nests. It is intended to accompany a manuscript submitted to Journal of the Royal Society Interface. The function allows for optional input so that the user can easily play with the parameters in the model and visualize the resulting dynamics. This function was used to generate all data reported in the paper.